+++ docs/fr_FR/changelog.md
@@ -3,6 +3,10 @@
 >**IMPORTANT**
 >
 >Pour rappel s'il n'y a pas d'information sur la mise à jour, c'est que celle-ci concerne uniquement de la mise à jour de documentation, de traduction ou de texte.
+# 13/04/2023
+- correction gestion validation token ima protect
+- correction fonction de prise de snapshot
+- nettoyage code
 
 # 05/04/2023
 - correction récupération de la liste des contacts ainsi que de la fonction arrêt de l'alarme